requirements.txt
pymongo
pandas 
numpy 
python-dotenv
PyYAML
dill
scikit-learn
-e .